Make the thing compile
After the rebase, make the thing compile.  I'm not really sure how I
feel about the VM *vm being _everywher_ but I'll continue and see if I
can make things better...
