+++ stdlib/_init.c
@@ -3,9 +3,9 @@
 
 void init_stdlib(VM *vm)
 {
-    VALUE obj_klass = bootstrapNativeClass("Object", NULL, NULL);
-    init_string(obj_klass);
-    init_object(obj_klass);
+    VALUE obj_klass = bootstrapNativeClass(vm, "Object", NULL, NULL);
+    init_string(vm, obj_klass);
+    init_object(vm, obj_klass);
     init_exception(vm);
     init_file(vm);
     init_iterable(vm);
